third_party/WebKit/LayoutTests/fast/dom/MutationObserver/mutate-during-delivery-expected.txt

Test that mutations during delivery do not interrupt delivery order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS order.length is 3
PASS order[0] is 1
PASS order[1] is 3
PASS order[2] is 2
PASS successfullyParsed is true

TEST COMPLETE
